tests refactoring + celery test
