+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
 DJANGO_SETTINGS_MODULE = autosaloon.settings
-addopts = -rP
+addopts = -rP --disable-warnings
