+++ tests/conftest.py
@@ -1,17 +1,14 @@
-from decimal import Decimal
 import pytest
 from pytest_factoryboy import register
 from django.db import connections
-from rest_framework.test import APIClient, RequestsClient
-
-from django.contrib.auth.models import User
+from rest_framework.test import APIClient
 
 import psycopg2
 from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
 
 from tests.factories import AutosaloonFactory, ProfileFactory, AutoFactory, UserFactory, DealerFactory, \
-	DealertosaloonhistoryFactory, DealercarsFactory, SaloontobuyerhistoryFactory, DealerdiscountFactory, \
-	SalooncarsFactory, OfferFactory, SaloondiscountFactory
+    DealertosaloonhistoryFactory, DealercarsFactory, SaloontobuyerhistoryFactory, DealerdiscountFactory, \
+    SalooncarsFactory, OfferFactory, SaloondiscountFactory
 
 register(AutosaloonFactory)
 register(UserFactory)
@@ -28,291 +25,37 @@ register(SaloondiscountFactory)
 
 
 def run_sql(sql):
-	conn = psycopg2.connect(database='postgres', user='Autosaloon')
-	conn.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)
-	cur = conn.cursor()
-	cur.execute(sql)
-	conn.close()
-	
-	
+    conn = psycopg2.connect(database='postgres', user='Autosaloon')
+    conn.set_isolation_level(ISOLATION_LEVEL_AUTOCOMMIT)
+    cur = conn.cursor()
+    cur.execute(sql)
+    conn.close()
+
+
 @pytest.fixture(scope='session')
 def django_db_setup():
-	from django.conf import settings
-	
-	#settings.DATABASES['default']['NAME'] = 'the_copied_db'
-	
-	#run_sql('DROP DATABASE IF EXISTS the_copied_db')
-	#run_sql('CREATE DATABASE the_copied_db TEMPLATE postgres')
-	
-	#yield
-	
-	#for connection in connections.all():
-	#	connection.close()
-		
-	#run_sql('DROP DATABASE the_copied_db')
+    from django.conf import settings
 
-	settings.DATABASES['default'] = {
-		'ENGINE': 'django.db.backends.postgresql',
-		'HOST': 'localhost',
-		'NAME': 'postgres',
-		'ATOMIC_REQUESTS': False,
-	}
+    #settings.DATABASES['default']['NAME'] = 'the_copied_db'
 
+    #run_sql('DROP DATABASE IF EXISTS the_copied_db')
+    #run_sql('CREATE DATABASE the_copied_db TEMPLATE postgres')
 
-@pytest.fixture
-def profile_and_car_for_offer(db, profile_factory, auto_factory):
-	profile = profile_factory.create()
-	car = auto_factory.create()
-	return profile.id, car.id
+    #yield
 
+    #for connection in connections.all():
+    #	connection.close()
 
-@pytest.fixture
-def const_discount_setup(db, auto_factory, autosaloon_factory, dealer_factory):
-	car = auto_factory.create()
-	saloon = autosaloon_factory.create()
-	dealer = dealer_factory.create()
-	return car, saloon, dealer
-	
+    #run_sql('DROP DATABASE the_copied_db')
 
-@pytest.fixture
-def best_dealer_setup(db, auto_factory, autosaloon_factory):
-	car = auto_factory.create()
-	saloon = autosaloon_factory.create()
-	return car, saloon
-	
-	
-@pytest.fixture
-def find_cars_and_dealers_setup(db, auto_factory, autosaloon_factory):
-	cars = []
-	for name in 'ABC':
-		car = auto_factory(model_name=name)
-		cars.append(car)
-	saloon = autosaloon_factory()
-	return saloon, cars
-
-
-@pytest.fixture
-def cars_by_popularity_setup(db, autosaloon_factory, auto_factory, profile_factory):
-	cars_dict = {}
-	for name in 'ABC':
-		car = auto_factory(model_name=name)
-		cars_dict[car.model_name] = car
-	saloon = autosaloon_factory()
-	profile = profile_factory()
-	return saloon, profile, cars_dict
-	
-	
-@pytest.fixture
-def look_for_better_discounts_setup(db, auto_factory, autosaloon_factory, dealercars_factory, dealer_factory):
-	car = auto_factory(model_name='TestCar')
-	saloon = autosaloon_factory()
-	
-	for i in 'ABC':
-		dealercars_factory(
-			dealer = dealer_factory(name=i),
-			car = car,
-			car_price = Decimal(10000))
-	
-	return car, saloon
-
-
-@pytest.fixture
-def saloon_buys_update_db_setup(db, auto_factory, dealercars_factory, autosaloon_factory):
-	car = auto_factory(model_name='TestCar')
-	dealer_rec = dealercars_factory(car=car)
-	saloon = autosaloon_factory()
-	return car, dealer_rec, saloon
-
-	
-@pytest.fixture
-def saloon_buys_car_setup(db, dealer_factory, auto_factory, dealercars_factory, autosaloon_factory, salooncars_factory):
-	cars = []
-	dealer = dealer_factory()
-	for name in 'ABC':
-		car = auto_factory(model_name=name)
-		if name == 'A':
-			car_in_saloon = car
-		if name == 'B':
-			dealercars_factory(car=car, dealer=dealer, car_price = Decimal(30000))
-		else:
-			dealercars_factory(car=car, dealer=dealer)
-		cars.append(car.id)
-	saloon = autosaloon_factory(balance=Decimal(20000))
-	salooncars_factory(saloon=saloon, car=car_in_saloon, quantity=3)
-	
-	return cars, saloon
-
-
-@pytest.fixture
-def find_best_offer_setup(db, auto_factory, offer_factory):
-	car = auto_factory()
-	offer = offer_factory(car_model=car)
-	return offer
-
-
-@pytest.fixture
-def customer_buys_update_db_setup(db, auto_factory, offer_factory, salooncars_factory):
-	car = auto_factory()
-	offer = offer_factory(car_model=car)
-	saloon_offer = salooncars_factory(car=car)
-	best_offer = [Decimal(8000), saloon_offer]
-	return offer, best_offer
-
-
-@pytest.fixture
-def customer_buys_car_setup(db, offer_factory):
-	offer = offer_factory()
-	return offer
+    settings.DATABASES['default'] = {
+        'ENGINE': 'django.db.backends.postgresql',
+        'HOST': 'localhost',
+        'NAME': 'postgres',
+        'ATOMIC_REQUESTS': False,
+    }
 
 
 @pytest.fixture
 def client():
-	return APIClient()
-
-# Autosaloon api fixtures
-
-@pytest.fixture
-def setup_cars_saloon_patch(db, user_factory, autosaloon_factory):
-	user = user_factory()
-	wrong_user = user_factory(username="wrong")
-	saloon = autosaloon_factory()
-	autosaloon_factory(name="wrong_saloon")
-	saloon.admin = user
-	saloon.save()
-	return saloon, wrong_user
-
-
-@pytest.fixture
-def setup_cars_saloon_cars_list(db, user_factory, auto_factory, autosaloon_factory, salooncars_factory):
-	admin_user = user_factory()
-	wrong_user = user_factory(username="wrong")
-	cars = [auto_factory(model_name=model) for model in 'XYZ']
-	for name in 'ABC':
-		saloon = autosaloon_factory(name=name)
-		for car in cars:
-			saloon_car = salooncars_factory(saloon=saloon, car=car)
-	saloon.admin = admin_user
-	saloon.save()
-	return saloon, wrong_user, saloon_car
-
-
-@pytest.fixture
-def setup_cars_saloon_cars_delete(db, user_factory, salooncars_factory):
-	admin_user = user_factory()
-	wrong_user = user_factory(username="wrong")
-	saloon_car = salooncars_factory()
-	saloon = saloon_car.saloon
-	saloon.admin = admin_user
-	saloon.save()
-	return admin_user, wrong_user, saloon_car
-
-
-@pytest.fixture
-def setup_cars_saloon_cars_add_discount(db, user_factory, salooncars_factory, saloondiscount_factory, autosaloon_factory):
-	admin_user = user_factory()
-	wrong_user = user_factory(username="wrong")
-	saloon_car = salooncars_factory()
-	saloon = saloon_car.saloon
-	saloon.admin = admin_user
-	saloon.save()
-	discount = saloondiscount_factory(seller=saloon)
-	wrong_saloon = autosaloon_factory(name="wrong")
-	wrong_discount = saloondiscount_factory(seller=wrong_saloon)
-	return admin_user, wrong_user, saloon_car, discount, wrong_discount
-
-
-@pytest.fixture
-def setup_cars_saloon_cars_remove_discount(db, user_factory, saloondiscount_factory, salooncars_factory):
-	admin_user = user_factory()
-	wrong_user = user_factory(username="wrong_user")
-	discount = saloondiscount_factory()
-	wrong_discount = saloondiscount_factory(seller__name="wrong_saloon")
-	saloon = discount.seller
-	saloon.admin = admin_user
-	saloon.save()
-	saloon_car = salooncars_factory(saloon=saloon)
-	saloon_car.car_discount.add(discount.id)
-	return admin_user, wrong_user, saloon_car, discount, wrong_discount
-
-
-# user api fixtures
-
-@pytest.fixture
-def setup_cars_autos(db, auto_factory):
-	for i in 'ABC':
-		car = auto_factory(model_name=i)
-	return car.id
-
-
-@pytest.fixture
-def setup_cars_saloon_list(db, user_factory, autosaloon_factory):
-	user = user_factory()
-	wrong_user = user_factory(username="wrong")
-	for i in 'ABC':
-		saloon = autosaloon_factory(name=i)
-	saloon.admin = user
-	saloon.save()
-	return saloon, wrong_user
-
-
-@pytest.fixture
-def setup_trading_profile_list(db, user_factory, profile_factory):
-	no_profile_user = user_factory(username="no_profile")
-	profile = profile_factory()
-	wrong_profile = profile_factory(user__username='wrong_user')
-	return profile, wrong_profile, no_profile_user
-
-
-@pytest.fixture
-def setup_trading_profile_post(db, user_factory, profile_factory):
-	no_profile_user = user_factory(username="no_profile")
-	profile = profile_factory()
-	return no_profile_user, profile
-
-
-@pytest.fixture
-def setup_trading_make_offer(db, profile_factory, salooncars_factory, user_factory):
-	profile = profile_factory()
-	no_profile_user = user_factory(username="no_profile")
-	record = salooncars_factory()
-	return profile, record, no_profile_user
-
-
-@pytest.fixture
-def setup_trading_saloon_discounts_list_retrieve(db, autosaloon_factory, salooncars_factory, user_factory, saloondiscount_factory):
-	saloon = autosaloon_factory(name='testsaloon')
-	salooncar_rec = salooncars_factory(saloon=saloon)
-
-	admin = user_factory()
-	wrong_user = user_factory(username="wrong_user")
-	saloon.admin = admin
-	saloon.save()
-
-	discount = saloondiscount_factory(seller=saloon)
-	salooncar_rec.car_discount.add(discount.id)
-
-	saloondiscount_factory()
-
-	return admin, wrong_user, discount
-
-
-@pytest.fixture
-def setup_trading_saloon_discounts_post(db, user_factory, salooncars_factory):
-	admin = user_factory(username='saloon_admin')
-	wrong_user = user_factory()
-	salooncar_rec = salooncars_factory()
-	saloon = salooncar_rec.saloon
-	saloon.admin = admin
-	saloon.save()
-	return saloon, wrong_user, salooncar_rec
-
-
-@pytest.fixture
-def setup_trading_saloon_discounts_delete(db, user_factory, saloondiscount_factory):
-	discount = saloondiscount_factory()
-	saloon = discount.seller
-	admin = user_factory(username='saloon_admin')
-	wrong_user = user_factory()
-	saloon.admin = admin
-	saloon.save()
-	return admin, wrong_user, discount
+    return APIClient()
