+++ tests/endpoints/saloon/test_saloon_endpoints.py
@@ -2,34 +2,10 @@ import pytest
 import datetime
 
 from cars.models import AutoSaloon, SaloonCars
+from tests.endpoints.utils import client_login, make_endpoint
 from trading.models import SaloonDiscount
 
 pytestmark = pytest.mark.django_db
-
-
-def client_login(client, user_type, correct_user, wrong_user):
-
-    """Authenticates correct or wrong users for tests"""
-
-    if user_type == 'correct':
-        client.force_authenticate(user=correct_user)
-    elif user_type == 'wrong':
-        client.force_authenticate(user=wrong_user)
-    return client
-
-
-def make_endpoint(base_url, endpoint_type, corr_value, wrong_value=None, wrong_id_value=None):
-
-    """Adds correct/wrong/non-existent ids to the end of url"""
-
-    url = base_url
-    if endpoint_type == "correct":
-        url += f'{corr_value}/'
-    elif endpoint_type == "wrong":
-        url += f'{wrong_value}/'
-    elif endpoint_type == "wrong_id":
-        url += f'{wrong_id_value}/'
-    return url
 
 
 @pytest.mark.parametrize(
